add NewChatFrom for refactoring
